requirements/dev.txt
-r base.txt
django-extensions
black
isort
flake8
django-debug-toolbar